일시정지 기능, 장면 전환 로딩 화면, 백그라운드 이미지 개선, 이미지 오브젝트 효과 추가
1. 게이트를 통해 장면 전환을 할 때 오브젝트들의 Update를 제어하여 일시 정지하는 기능 구현
2. 장면을 전환할 때 캐릭터가 달리고 있는 모션이 출력되는 애니메이션을 그리도록 함
3. 배경 이미지가 카메라의 이동에 따라 DepthLevel을 부여하여 깊이감에 따라 각각 움직이도록 만듬.
4. 이미지 오브젝트에 깜박이는 효과 추가
5. 대시기능 세컨드대시가 다시 충전 안되는 이슈 -> 해결중

+++ WinAPI_Skul/CUI.cpp
@@ -76,10 +76,10 @@ void CUI::Render(HDC _hDC)
 	if (m_bLbtnDown)
 	{
 		RENDER->RenderFillRectangle(
-			(int)(vPos.x),
-			(int)(vPos.y),
-			(int)(vPos.x + vScale.x),
-			(int)(vPos.y + vScale.y));
+			vPos.x,
+			vPos.y,
+			vPos.x + vScale.x,
+			vPos.y + vScale.y);
 			
 	}
 	else
